Fix bug in the rounding logic
Note, this is a "fix" in that it makes the logic correct for finding the
shortest float in the range if one exists. However, it appears to break
the output on one single float (out of all ~4 billion floats, just 1
fails). The one that fails is 7.03853131E-26, which falsely gets
converted to "7.038531E-26". I'll need to look into this. For now I am
going to push this anyway, because it fixes the general case.