Bump valid interval limits to fix error case
This commit offsets the high and low interval bounds for numerical
stability at the edges. Doing this slightly narrows the range of valid
outputs. In some cases it may mean we miss the shortest possible
representation, but that is preferable to outputting something outside
the valid range.

With this change, we no longer output an incorrect answer for any
number.